+++ livestream/views.py
@@ -1,13 +1,11 @@
 from django.shortcuts import render
-from django.http import HttpResponse
 from django.views.generic.detail import SingleObjectMixin
-from django.views import generic, View
+from django.views import generic
 from rest_framework.response import Response
-from rest_framework import status, viewsets, renderers
-from rest_framework.renderers import TemplateHTMLRenderer
+from rest_framework import status, viewsets
 
-from .serializers import *
-from .models import Appeal
+from .serializers import AppealSerializer, ApprovalRequestSerializer
+from .models import Appeal, ApprovalRequest
 
 from aLive.settings import OPENTOK_API, OPENTOK_SECRET
 
@@ -112,42 +110,65 @@ class AppealDetailView(generic.DetailView):
         return self.render_to_response(self.context)
 
 
-# UserViewSet
+class ApprovalRequestViewSet(viewsets.ModelViewSet):
+    queryset = ApprovalRequest.objects.all()
+    serializer_class = ApprovalRequestSerializer
 
+    def create(self, request, *args, **kwargs):
+        # when user presses "HELP" button
+        # an ApprovalRequest instance gets created,
+        # UNLESS it already exists
+        appeal_instance = self.request.data['appeal']
+        if ApprovalRequest.objects.filter(
+            appeal=appeal_instance,
+                helper=self.request.user).exists():
+            # WARNING: if request gets rejected (is_accepted holds false)
+            # return message will still be 'pending approval...'
+            return Response({'return': 'pending approval...'})
 
+        # owner should NOT BE ABLE TO create approvalrequess
+        # for their OWN appeals, that's stupid
+        if appeal_instance.owner == self.request.user:
+            return Response({'return': "action impossible"})
 
-# class ClientTokenViewSet(viewsets.ModelViewSet):
-#     '''
-#         To change:
-#             Tokens should not be stored or reused
-#     '''
-#     queryset = ClientToken.objects.all()
-#     serializer_class = ClientTokenSerializer
+        serializer = self.get_serializer(data=request.data)
+        serializer.is_valid(raise_exception=True)
+        self.perform_create(serializer)
+        headers = self.get_success_headers(serializer.data)
+        return Response(serializer.data, status=status.HTTP_201_CREATED,
+                        headers=headers)
 
-#     def create(self, request):
-#         ret = {'return': 'token creation failed'}
-#         req = request.data
-#         # get session id from db
-#         this_session = Appeal.objects.get(id=req['session'])
-#         # generate token
-#         token = opentok.generate_token(this_session.session_id)
-#         # get user
-#         user = User.objects.get(id=req['user'])
-#         new_token = ClientToken(token_id=token,
-#                                 # fixed wrong parameter thingy
-#                                 session=this_session,
-#                                 user=user)
-#         # check if new token was successfully created
-#         # and user does not have existing token
-#         print(ClientTokenSerializer(new_token).data)
-#         if new_token:
-#             print(new_token)
-#             # Value error here
-#             new_token.save()
-#             # VALUE ERROR SOLVED
-#             ret = ClientTokenSerializer(new_token).data
+    def list(self, request, *args, **kwargs):
+        # display list of all ApprovalRequests
+        # should only display ApprovalRequests that have
+        # not been rejected (is_accepted holds null)
+        queryset = ApprovalRequest.objects.exclude(is_approved=False)
 
-#         return Response(ret)
+        page = self.paginate_queryset(queryset)
+        if page is not None:
+            serializer = self.get_serializer(page, many=True)
+            return self.get_paginated_response(serializer.data)
+
+        serializer = self.get_serializer(queryset, many=True)
+        return Response(serializer.data)
+
+    def partial_update(self, request, *args, **kwargs):
+        kwargs['partial'] = True
+        return self.update(request, *args, **kwargs)
+
+    def destroy(self, request, *args, **kwargs):
+        # when approval request gets rejected it gets deleted from the db
+        # when user revokes approval request it gets deleted from the db
+        # when a request gets accepted other instances
+        # from the same appeal should be deleted
+        instance = self.get_object()
+        self.perform_destroy(instance)
+        return Response(status=status.HTTP_204_NO_CONTENT)
+
+    def retrieve(self, request, *args, **kwargs):
+        instance = self.get_object()
+        serializer = self.get_serializer(instance)
+        return Response(serializer.data)
 
 
 class IndexView(generic.ListView):
